+++ sickbeard/name_parser/parser.py
@@ -135,24 +135,40 @@ class NameParser(object):
                     result.season_number = tmp_season
                     result.score += 1
 
-                if 'ep_num' in named_groups:
-                    ep_num = self._convert_number(match.group('ep_num'))
-                    if 'extra_ep_num' in named_groups and match.group('extra_ep_num'):
-                        result.episode_numbers = range(ep_num, self._convert_number(match.group('extra_ep_num')) + 1)
-                        result.score += 1
+                def _process_epnum(captures, capture_names, grp_name, extra_grp_name, ep_numbers, parse_result):
+                    ep_num = self._convert_number(captures.group(grp_name))
+                    extra_grp_name = 'extra_%s' % extra_grp_name
+                    ep_numbers = '%sepisode_numbers' % ep_numbers
+                    if extra_grp_name in capture_names and captures.group(extra_grp_name):
+                        try:
+                            if hasattr(self.showObj, 'getEpisode'):
+                                ep = self.showObj.getEpisode(parse_result.season_number, ep_num)
+                            else:
+                                tmp_show = helpers.get_show(parse_result.series_name, True, False)
+                                if tmp_show and hasattr(tmp_show, 'getEpisode'):
+                                    ep = tmp_show.getEpisode(parse_result.season_number, ep_num)
+                                else:
+                                    ep = None
+                        except:
+                            ep = None
+                        en = ep and ep.name and re.match(r'^\W*(\d+)', ep.name) or None
+                        es = en and en.group(1) or None
+
+                        extra_ep_num = self._convert_number(captures.group(extra_grp_name))
+                        parse_result.__dict__[ep_numbers] = range(ep_num, extra_ep_num + 1) if not (
+                            ep and es and es != captures.group(extra_grp_name)) and (
+                            0 < extra_ep_num - ep_num < 10) else [ep_num]
+                        parse_result.score += 1
                     else:
-                        result.episode_numbers = [ep_num]
-                    result.score += 1
+                        parse_result.__dict__[ep_numbers] = [ep_num]
+                    parse_result.score += 1
+                    return parse_result
+
+                if 'ep_num' in named_groups:
+                    result = _process_epnum(match, named_groups, 'ep_num', 'ep_num', '', result)
 
                 if 'ep_ab_num' in named_groups:
-                    ep_ab_num = self._convert_number(match.group('ep_ab_num'))
-                    if 'extra_ab_ep_num' in named_groups and match.group('extra_ab_ep_num'):
-                        result.ab_episode_numbers = range(ep_ab_num,
-                                                          self._convert_number(match.group('extra_ab_ep_num')) + 1)
-                        result.score += 1
-                    else:
-                        result.ab_episode_numbers = [ep_ab_num]
-                    result.score += 1
+                    result = _process_epnum(match, named_groups, 'ep_ab_num', 'ab_ep_num', 'ab_', result)
 
                 if 'air_year' in named_groups and 'air_month' in named_groups and 'air_day' in named_groups:
                     year = int(match.group('air_year'))