+++ CHANGES.md
@@ -38,6 +38,7 @@
 * Add Emby notifier to config/Notifications
 * Fix alternative unicode show names from breaking search
 * Change show update, set shows with newly added airdate or existing episodes with future or never dates, to "Wanted"
+* Fix rare NameParser case where numeric episode name was parsed as episode number
 
 
 ### 0.11.6 (2016-02-18 23:10:00 UTC)